chore: move override typings and files
