+++ packages/core/src/components/sections/BannerText/Overrides.tsx
@@ -3,9 +3,9 @@ import {
   BannerTextContent as UIBannerTextContent,
 } from '@faststore/ui'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/BannerText'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const { BannerText, BannerTextContent } = getSectionOverrides(
   {
