+++ packages/core/src/components/sections/EmptyState/Overrides.tsx
@@ -1,8 +1,8 @@
 import { EmptyState as UIEmptyState } from '@faststore/ui'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/EmptyState'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const { EmptyState } = getSectionOverrides(
   {
