+++ packages/core/src/components/sections/Hero/Overrides.tsx
@@ -4,9 +4,9 @@ import {
   HeroHeader as UIHeroHeader,
 } from '@faststore/ui'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/Hero'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const { Hero, HeroImage, HeroHeader } = getSectionOverrides(
   {
