+++ packages/core/src/components/sections/Navbar/Overrides.tsx
@@ -12,9 +12,9 @@ import {
   IconButton as UIIconButton,
 } from '@faststore/ui'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/Navbar'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const {
   Navbar,
