+++ packages/core/src/components/sections/Newsletter/Overrides.tsx
@@ -5,9 +5,9 @@ import {
 } from '@faststore/ui'
 import { NewsletterAddendum } from 'src/components/ui/Newsletter/NewsletterAddendum'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/Newsletter'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const {
   ToastIconSuccess,
