+++ packages/core/src/components/sections/ProductShelf/Overrides.tsx
@@ -3,9 +3,9 @@ import { ProductShelf as UIProductShelf } from '@faststore/ui'
 import ProductCard from 'src/components/product/ProductCard'
 import Carousel from 'src/components/ui/Carousel'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/ProductShelf'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const { ProductShelf, __experimentalCarousel, __experimentalProductCard } =
   getSectionOverrides(
