+++ packages/core/src/components/sections/RegionBar/Overrides.tsx
@@ -1,8 +1,8 @@
 import { RegionBar as UIRegionBar, Icon as UIIcon } from '@faststore/ui'
 
-import { getSectionOverrides } from 'src/utils/overrides'
+import { getSectionOverrides } from 'src/sdk/overrides/overrides'
 import { override } from 'src/customizations/src/components/overrides/RegionBar'
-import type { SectionOverrideDefinition } from 'src/typings/overrides'
+import type { SectionOverrideDefinition } from 'src/typings/overridesDefinition'
 
 const { RegionBar, LocationIcon, ButtonIcon } = getSectionOverrides(
   {
